IT WORKS
Sorry, too far gone for a real commit message... Need to test it with
different # of cores to see if parallelization is actually happening. Also,
pretty sure it won't solve the snowflake puzzles (same problem my Python
prototype had - and actually that algorithm was a bit more sophisticated).
For now I need to nap.

Signed-off-by: ParceledTongue <zach@palumbo.com>